pyproject.toml
[tool.poetry]
name = "todo"
version = "1.0.0"
description = "A simple bot to keep track of things you need to do"
authors = ["vcokltfre <vcokltfre@gmail.com>"]
license = "MIT"
repository = "https://github.com/opendiscordbots/todobot"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.5.3"
taskipy = "^1.8.2"
python-dotenv = "^0.19.0"
disnake = "^2.2.1"
libodb = "^1.0.1"

[tool.poetry.dev-dependencies]

[tool.taskipy.tasks]
start = "python3 -m src"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
